fix: big change - decoupled all molstar related functions into multiple files
